fix(platform): reject duplicate customer add on the reachable create path (#1993)
The customer fix previously lived in createCustomerPublic, a dead helper imported only by its own test and never registered as a Convex endpoint. Move duplicate detection (email + externalId) into create_customer.ts, the helper backing internal_mutations.createCustomer, which is the path REST POST /api/v1/customers, the agent customer tool, and the workflow customer action all run. Normalize email to match the dup index and keep findOrCreateCustomer idempotent. Drop the dead helper, its test, and the now-unreachable CUSTOMER_EMAIL_REQUIRED 400 mapping (the reachable path allows email-optional adds). Add create_customer.test.ts covering both duplicate codes and the happy paths.
