+++ services/platform/convex/customers/create_customer.test.ts
@@ -2,7 +2,7 @@ import { convexTest } from 'convex-test';
 import { describe, expect, it } from 'vitest';
 
 import schema from '../schema';
-import { createCustomerPublic } from './create_customer_public';
+import { createCustomer } from './create_customer';
 
 // convex-test module map, keyed relative to the convex/ root. This file lives at
 // convex/customers/, so the glob reaches the root via ../ and keys are
@@ -48,10 +48,13 @@ function codeOf(err: unknown): string | undefined {
   return typeof candidate === 'string' ? candidate : undefined;
 }
 
-describe('createCustomerPublic duplicate guard (#1993)', () => {
-  // Regression: the duplicate-add rejections must carry structured ConvexError
-  // codes instead of raw `Error`s, which Convex redacts to "Server Error" and
-  // forces downstream string-matching of the message.
+// Regression for #1993: the reachable customer-add path (REST
+// `POST /api/v1/customers`, the agent customer tool, and the workflow customer
+// action all run `internal.customers.internal_mutations.createCustomer`, which
+// delegates to this helper) must reject duplicate adds with a structured
+// ConvexError code instead of a raw `Error` (which Convex redacts to "Server
+// Error" in prod and forces downstream message string-matching).
+describe('createCustomer duplicate guard (#1993)', () => {
   it('throws ConvexError({ code: CUSTOMER_DUPLICATE_EMAIL }) on an email collision', async () => {
     const t = convexTest(schema, modules);
 
@@ -67,7 +70,7 @@ describe('createCustomerPublic duplicate guard (#1993)', () => {
     let code: string | undefined;
     try {
       await t.run((ctx) =>
-        createCustomerPublic(ctx, {
+        createCustomer(ctx, {
           organizationId: ORG,
           // Mixed case + whitespace normalizes to the existing email.
           email: '  Dup@Example.com ',
@@ -98,7 +101,7 @@ describe('createCustomerPublic duplicate guard (#1993)', () => {
     let code: string | undefined;
     try {
       await t.run((ctx) =>
-        createCustomerPublic(ctx, {
+        createCustomer(ctx, {
           organizationId: ORG,
           email: 'fresh@example.com',
           externalId: 'ext-7',
@@ -113,31 +116,11 @@ describe('createCustomerPublic duplicate guard (#1993)', () => {
     expect(code).toBe('CUSTOMER_DUPLICATE_EXTERNAL_ID');
   });
 
-  it('throws ConvexError({ code: CUSTOMER_EMAIL_REQUIRED }) on an empty email', async () => {
+  it('inserts the customer (normalizing email) when no collision exists', async () => {
     const t = convexTest(schema, modules);
 
-    let code: string | undefined;
-    try {
-      await t.run((ctx) =>
-        createCustomerPublic(ctx, {
-          organizationId: ORG,
-          email: '   ',
-          status: 'active',
-          source: 'manual_import',
-        }),
-      );
-    } catch (err) {
-      code = codeOf(err);
-    }
-
-    expect(code).toBe('CUSTOMER_EMAIL_REQUIRED');
-  });
-
-  it('inserts the customer when no collision exists', async () => {
-    const t = convexTest(schema, modules);
-
-    const id = await t.run((ctx) =>
-      createCustomerPublic(ctx, {
+    const result = await t.run((ctx) =>
+      createCustomer(ctx, {
         organizationId: ORG,
         email: 'New@Example.com',
         status: 'active',
@@ -145,7 +128,26 @@ describe('createCustomerPublic duplicate guard (#1993)', () => {
       }),
     );
 
-    const row = await t.run((ctx) => ctx.db.get(id));
+    const row = await t.run((ctx) => ctx.db.get(result.customerId));
+    expect(result.success).toBe(true);
     expect(row?.email).toBe('new@example.com');
+  });
+
+  it('allows an email-less customer add (agents may create with only a name)', async () => {
+    const t = convexTest(schema, modules);
+
+    const result = await t.run((ctx) =>
+      createCustomer(ctx, {
+        organizationId: ORG,
+        name: 'No Email Co',
+        status: 'active',
+        source: 'manual_import',
+      }),
+    );
+
+    const row = await t.run((ctx) => ctx.db.get(result.customerId));
+    expect(result.success).toBe(true);
+    expect(row?.name).toBe('No Email Co');
+    expect(row?.email).toBeUndefined();
   });
 });
