+++ services/platform/convex/lib/rest/helpers.ts
@@ -338,7 +338,6 @@ function httpStatusForConvexCode(code: string | undefined): number {
     case 'validation':
     case 'EMAIL_REQUIRED':
     case 'MISSING_FILTER':
-    case 'CUSTOMER_EMAIL_REQUIRED':
       return 400;
     // Duplicate-add rejections: a conflicting row already exists. Map to 409
     // so REST clients get an actionable conflict instead of an opaque 500.
